Hanging
Now, There is something that I can't seem to fix. and that is the Hanging after the explorer restart. I believe the cause of this is the 21H1 Update for Windows 10. But if it hangs and the explorer comes back on, (Icons, Recycle Bin, etc.) Just close it because it's already done it's job.